refactor(agent): share _last_assistant_index with preflight estimator
Addresses AI-review feedback on #86055:
- Reuse the existing newest-assistant scan (moved to model_metadata,
  the dependency-lowest owner already imported by ContextCompressor)
  instead of duplicating the loop inline. Preflight and tail-budget
  walks now share one definition of "newest" and cannot drift.
- Skip the dict rebuild for older assistant rows that carry no
  reasoning / reasoning_content keys at all, avoiding an allocation
  in the common case.

Verified: 92 tests pass (test_model_metadata + test_replay_budget_accounting);
new estimator output is numerically identical to the inline-loop version
across 200 randomized transcripts.
